updated drop and pickup services

+++ src/services/pickup-address.service.ts
@@ -5,24 +5,29 @@ import { PickupProvider, PrismaClient } from '@prisma/client';
 import { PickupAddressInput } from '@/graphql-type/args/pickup-address.input';
 import { nanoid } from 'nanoid';
 import { PickupAddressType } from '@/graphql-type/typedefs/pickup-address.type';
+import OrderService from './order.service';
+import { badUserInputException } from '@/utils/exceptions.util';
+import { USER_ERROR_KEYS } from '@/constants';
 
 @Service()
 export default class PickupAddressService {
   constructor(
     @Inject('prisma') private prisma: PrismaClient,
     @Inject('logger') private logger: typeof LoggerInstance,
+    private orderService: OrderService,
     @EventDispatcher() private eventDispatcher: EventDispatcherInterface,
-  ) {}
+  ) { }
 
-  public async getPickupAddresses({ userId} : { userId: string }) {
+  public async getPickupAddresses({ userId }: { userId: string }) {
     return this.prisma.pickupAddress.findMany({
       where: {
+        isDeleted: false,
         userId,
       }
     }) as Promise<PickupAddressType[]>
   }
 
-  public async addPickupAddress({ input, userId} : { input: PickupAddressInput; userId: string }) {
+  public async addPickupAddress({ input, userId }: { input: PickupAddressInput; userId: string }) {
     return this.prisma.pickupAddress.create({
       data: {
         pickupId: nanoid(),
@@ -32,30 +37,71 @@ export default class PickupAddressService {
     }) as Promise<PickupAddressType>
   }
 
-  public async deletePickupAddress({ id, userId } : { id: string, userId: string }) {
-    // TODO: delete pickup address if not used in any order, otherwise mark it as inactive
-    return this.prisma.pickupAddress.delete({
-      where: {
-        id,
-        userId
+  public async deletePickupAddress({ id, userId }: { id: string, userId: string }) {
+    try {
+      const orderCount = await this.orderService.getOrderCount({
+        userId,
+        where: {
+          pickupId: id
+        }
+      })
+      if (orderCount > 0) {
+        await this.prisma.pickupAddress.update({
+          where: {
+            id,
+            userId
+          },
+          data: {
+            isDeleted: false
+          }
+        })
+      } else {
+        await this.prisma.pickupAddress.delete({
+          where: {
+            id,
+            userId
+          },
+        })
       }
-    })
+    } catch (error) {
+      this.logger.error(error);
+      throw badUserInputException(USER_ERROR_KEYS.INVALID_REQUEST);
+    }
   }
 
-  public async editPickupAddress({ input, id, userId} : { input: PickupAddressInput; id: string, userId: string }) {
-    // TODO: update pickup address if not used in any order, otherwise throw error 
-    return this.prisma.pickupAddress.update({
-      where: {
-        id,
-        userId
-      },
-      data: {
-        ...input,
-      }
-    })
+  public async editPickupAddress({ input, userId }: { input: { id: string; updatedData: PickupAddressInput }, userId: string }) {
+    const { id, updatedData } = input
+    try {
+      const orderCount = await this.orderService.getOrderCount({
+        userId,
+        where: {
+          pickupId: id
+        }
+      })
+      const isOrderExist = orderCount > 0;
+      let data = await this.prisma.pickupAddress.update({
+          where: {
+            id,
+            userId
+          },
+          data: isOrderExist ? {
+            isDeleted: false
+          } : updatedData
+        }) as PickupAddressType
+       if(isOrderExist) {
+         data = await this.addPickupAddress({ input: updatedData, userId }) as PickupAddressType;
+       }
+       return {
+        isNew: isOrderExist,
+        data
+       }
+    } catch (error) {
+      this.logger.error(error);
+      throw badUserInputException(USER_ERROR_KEYS.INVALID_REQUEST);
+    }
   }
 
-  public async createPickupProvider({ data } : { data: PickupProvider }) {
+  public async createPickupProvider({ data }: { data: PickupProvider }) {
     return this.prisma.pickupProvider.create({
       data: {
         ...data
@@ -63,14 +109,4 @@ export default class PickupAddressService {
     })
   }
 
-  public async updatePickupProvider({ updatedData, id } : { updatedData: PickupProvider; id: string }) {
-    return this.prisma.pickupAddress.update({
-      where: {
-        id,
-      },
-      data: {
-        ...updatedData,
-      }
-    })
-  }
 }
